Moved levels to a separate file.
