+++ index.html
@@ -50,7 +50,7 @@
     
     </div>
 
-    <script src="./script.js"></script>
+    <script type="module" src="./script.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/tone/14.8.26/Tone.js"></script>
 </body>
 </html>
