+++ levels.js
@@ -1,0 +1,36 @@
+let levels = [
+    // duration in seconds
+    {
+        name: "C triad",
+        notes: [
+            {letter: 'C', octave: '4', duration: 1},
+            {letter: 'E', octave: '4', duration: 1},
+            {letter: 'G', octave: '4', duration: 2}
+        ]
+    },
+
+    {
+        name: "Easy as 1-2-3",
+        notes: [
+            {letter: 'A', octave: '3', duration: 1},
+            {letter: 'B', octave: '3', duration: 1},
+            {letter: 'C', octave: '4', duration: 2}
+        ]
+    },
+
+    {
+        name: "Lead frog 3rds",
+        notes: [
+            {letter: 'C', octave: '4', duration: 0.5},
+            {letter: 'E', octave: '4', duration: 0.5},
+            {letter: 'D', octave: '4', duration: 0.5},
+            {letter: 'F', octave: '4', duration: 0.5},
+            {letter: 'E', octave: '4', duration: 0.5},
+            {letter: 'G', octave: '4', duration: 0.5},
+        ]
+    },
+
+    // More levels...
+];
+
+export { levels };
